refactor: Add DRY <ReturnToGameZoneButton/>
- Refactor shared UI logic into new reusable UI component
- Currently used in: Accessibility Statement, Reset Confirm, & Not Found (404) pages
